production_agent/requirements.txt
# Core Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
typing_extensions>=4.9.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.30
langchain-core>=0.1.0

# Redis
redis>=4.6.0

# Environment
python-dotenv>=1.0.0

# Development dependencies (optional, can be installed separately)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0